implement user authentication with JWT, including login, logout, and token refresh functionality

+++ src/utils/cloudinary.js
@@ -14,12 +14,17 @@ const uploadOnCloudinary = async (localFilePath)=>{
         const response = await cloudinary.uploader.upload(localFilePath,{
             resource_type:"auto"
         })
-        console.log("file is uploaded on cloudinary",response.url);
+        // console.log("file is uploaded on cloudinary",response.url);
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath)
+    console.log("CLOUDINARY ERROR:", error);
 
-        return null;
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
+    }
+
+    return null;
+
     }
 }
